requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
google-genai
aiofiles
